+++ config/splat.pspeu.stcat.yaml
@@ -61,7 +61,8 @@ segments:
       - [0x1A9C0, c, cat/st_common]
       - [0x1C308, c, cat_psp/func_pspeu_09253988]
       - [0x1C4B0, c, cat/prim_helpers]
-      - [0x1CD50, c, cat_psp/unk_1CD50]
+      - [0x1CD50, c, cat/en_gremlin]
+      - [0x1D8D0, c, cat_psp/unk_1D8D0]
       - [0x20A38, c, cat/en_thornweed_corpseweed]
       - [0x22740, c, cat/e_wereskeleton]
       - [0x23A98, c, cat_psp/unk_23A98]
@@ -83,7 +84,8 @@ segments:
       - [0x2CF78, .data, cat/e_room_fg]
       - [0x2D018, data]
       - [0x2D230, .data, cat/st_common]
-      - [0x2D430, data]
+      - [0x2D430, .data, cat/en_gremlin]
+      - [0x2D490, data]
       - [0x2D590, .data, cat/en_thornweed_corpseweed]
       - [0x2D660, .data, cat/e_wereskeleton]
       - [0x2D7C0, data]
