+++ config/splat.us.stcat.yaml
@@ -53,6 +53,8 @@ segments:
       - [0x24C8, .data, e_particles]
       - [0x2548, .data, e_room_fg]
       - [0x25D4, data]
+      - [0x278C, .data, en_gremlin]
+      - [0x27DC, data]
       - [0x2814, .data, e_wereskeleton]
       - [0x2990, .data, en_thornweed_corpseweed]
       - [0x2A48, data]
@@ -63,6 +65,8 @@ segments:
       - [0x36A5C, .rodata, e_stage_name]
       - [0x36A90, .rodata, unk_4A13C]
       - [0x36BB8, .rodata, unk_4EE88]
+      - [0x36CA8, .rodata, en_gremlin]
+      - [0x36CB4, .rodata, unk_52B3C]
       - [0x36D20, .rodata, e_wereskeleton]
       - [0x36DA4, .rodata, en_thornweed_corpseweed]
       - [0x36DE0, .rodata, unk_56C70]
@@ -85,6 +89,8 @@ segments:
       - [0x4EA6C, c, e_room_fg]
       - [0x4EB58, c, popup]
       - [0x4EE88, c, unk_4EE88]
+      - [0x522E4, c, en_gremlin]
+      - [0x52B3C, c, unk_52B3C]
       - [0x549F8, c, e_wereskeleton]
       - [0x55748, c, en_thornweed_corpseweed]
       - [0x56C70, c, unk_56C70]
