+++ config/symbols.pspeu.stcat.txt
@@ -88,6 +88,8 @@ g_EInitUnkId13 = 0x09269C58;
 g_EInitLockCamera = 0x09269C68;
 g_EInitCommon = 0x09269C78;
 g_EInitDamageNum = 0x09269C88;
+g_EInitGremlin = 0x09269DF8;
+g_EInitGremlinFire = 0x09269E08;
 g_EInitWereskeleton = 0x09269E38;
 D_us_80180A88 = 0x09269E48;
 g_EInitThornweed = 0x09269E58;
