+++ config/symbols.us.stcat.txt
@@ -13,6 +13,8 @@ g_EInitUnkId13 = 0x80181158;
 g_EInitLockCamera = 0x80181164;
 g_EInitCommon = 0x80181170;
 g_EInitDamageNum = 0x8018117C;
+g_EInitGremlin = 0x80181290;
+g_EInitGremlinFire = 0x801812A8;
 g_EInitWereskeleton = 0x801812CC;
 D_us_80180A88 = 0x801812D8;
 g_EInitThornweed = 0x801812E4;
