+++ src/st/cat/en_gremlin.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "cat.h"
+
+#include "../en_gremlin.h"
