+++ src/st/cat/unk_52B3C.c
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "cat.h"
+
+INCLUDE_ASM("st/cat/nonmatchings/unk_52B3C", func_us_801D2B3C);
+
+INCLUDE_ASM("st/cat/nonmatchings/unk_52B3C", func_us_801D3468);
+
+INCLUDE_ASM("st/cat/nonmatchings/unk_52B3C", func_us_801D3A50);
+
+INCLUDE_ASM("st/cat/nonmatchings/unk_52B3C", func_us_801D3C1C);
+
+INCLUDE_ASM("st/cat/nonmatchings/unk_52B3C", func_us_801D3D08);
+
+INCLUDE_ASM("st/cat/nonmatchings/unk_52B3C", func_us_801D4088);
+
+INCLUDE_ASM("st/cat/nonmatchings/unk_52B3C", func_us_801D4744);
