--- src/st/cat_psp/unk_1CD50.c
@@ -1,22 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../cat/cat.h"
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", EntityGremlin);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", EntityGremlinEffect);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", EntityGremlinFire);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", func_pspeu_09254F50);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", func_pspeu_09255558);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", func_pspeu_092566B8);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", func_pspeu_09257148);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", func_pspeu_092572D0);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", func_pspeu_09257850);
-
-INCLUDE_ASM("st/cat_psp/nonmatchings/cat_psp/unk_1CD50", func_pspeu_09257FD0);
